Make rmw_freertps libraries shared for now so they compile with ament
Move mega_msggen into the rosidl_typesupport_freertps module, change from file writing to string append
