+++ rosidl_typesupport_freertps_cpp/CMakeLists.txt
@@ -26,7 +26,7 @@ ament_export_include_directories(include include/${PROJECT_NAME}/impl)
 
 ament_python_install_package(${PROJECT_NAME})
 
-add_library(${PROJECT_NAME} src/identifier.cpp)
+add_library(${PROJECT_NAME} SHARED src/identifier.cpp)
 if(WIN32)
   target_compile_definitions(${PROJECT_NAME}
     PRIVATE "ROSIDL_TYPESUPPORT_OPENSPLICE_CPP_BUILDING_DLL")
