--- rosidl_typesupport_freertps_cpp/mega_msggen/generate_workspace.py
@@ -1,24 +1,0 @@
-#!/usr/bin/env python3
-import mega_msggen
-
-################################################3
-# begin global code awesomeness
-ament_prefix = os.environ['AMENT_PREFIX_PATH']
-ifaces_path = os.path.join(ament_prefix, 'share', 'ament_index', 'resource_index', 'rosidl_interfaces')
-share_path = os.path.join(ament_prefix, 'share')
-if not os.path.isdir(ifaces_path) or not os.path.isdir(share_path):
-  print("ament_index for rosidl_interfaces seems to be empty. Perhaps this workspace hasn't been built yet?")
-  sys.exit(1)
-
-msg_tree_root = os.path.join('build','msgs')
-if not os.path.exists(msg_tree_root):
-  os.makedirs(msg_tree_root)
-  os.makedirs(os.path.join(msg_tree_root,'src'))
-for pkg_name in os.listdir(ifaces_path):
-  full_path = os.path.join(ifaces_path, pkg_name)
-  pkg_output_path = os.path.join(msg_tree_root, pkg_name)
-  pkg_share_path = os.path.join(share_path, pkg_name)
-  if not os.path.exists(pkg_output_path):
-    os.makedirs(pkg_output_path)
-  with open(full_path) as f:
-    write_generated_code_to_file(f)
